refactor: extract fetchMultipleOffers in a new file
